full re-write of IMU library... derived from i2cdev library
The purpose of this change was to deal with the IMU error which has been interfering with the output measurements. This new library employs better filtration techniques and also represents the IMU outputs in quaternion form. This has the advantage of preventing gimbal lock, a common problem with the use of euler angles.

All rights attributed to the rightful owners. See more details: https://github.com/jrowberg/i2cdevlib